Multiple bug fix

+++ Wetzelplaats/Wetzelplaats-war/src/java/nl/wetzel/servlets/AdvertisementDeleteServlet.java
@@ -8,6 +8,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import nl.wetzel.entities.Advertisement;
+import nl.wetzel.entities.User;
+import nl.wetzel.entities.User.UserType;
 import nl.wetzel.facades.AdvertisementFacadeLocal;
 import nl.wetzel.facades.BidFacadeLocal;
 
@@ -19,30 +21,56 @@ public class AdvertisementDeleteServlet extends HttpServlet {
     @EJB
     private BidFacadeLocal bidFacade;
 
+    // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
+    /**
+     * Handles the HTTP
+     * <code>GET</code> method.
+     *
+     * @param request servlet request
+     * @param response servlet response
+     * @throws ServletException if a servlet-specific error occurs
+     * @throws IOException if an I/O error occurs
+     */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        int deleteId = Integer.parseInt(request.getParameter("did"));
-        int admin = Integer.parseInt(request.getParameter("admin"));
-
-        Advertisement ad = this.advertisementFacade.find(Integer.valueOf(deleteId));
-        this.bidFacade.deleteByAdId(ad);
-
-        this.advertisementFacade.deleteById(deleteId);
-
-        if (admin == 0) {
-            response.sendRedirect("/Wetzelplaats-war/account");
-        } else {
-            response.sendRedirect("/Wetzelplaats-war/admin");
+            throws ServletException, IOException 
+    {
+        if (request.getParameter("did") == null) {
+            response.sendRedirect("/Wetzelplaats-war/index");
+            return;
         }
+        
+        User user = (User) request.getSession().getAttribute("user");
+        int deleteId = Integer.parseInt(request.getParameter("did"));      
+        Advertisement ad = this.advertisementFacade.find(Integer.valueOf(deleteId));
+        
+        if(ad.getUserId().getId() == user.getId())
+        {
+            this.bidFacade.deleteByAdId(ad);
+
+            this.advertisementFacade.deleteById(ad);
+
+            if (user.getUserType() == UserType.Admin) 
+            {
+                response.sendRedirect("/Wetzelplaats-war/account");
+            } else {
+                response.sendRedirect("/Wetzelplaats-war/admin");
+            }
+        }
+        
+        else
+        {
+            response.sendRedirect("/Wetzelplaats-war/MagicWord");
+        } 
     }
 
-    /**
+       /**
+     * Returns a short description of the servlet.
      *
-     * @return
+     * @return a String containing servlet description
      */
     @Override
     public String getServletInfo() {
-        return "Short description";
+        return "Handles advertisement deletion";
     }
 }
